[MIG] Migrate printer_tray to v9.0
