+++ printer_tray/models/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2013-2014 Camptocamp (<http://www.camptocamp.com>)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from . import ir_actions_report_xml
+from . import printing_tray
+from . import printing_printer
+from . import res_users
+from . import printing_report_xml_action
