+++ printer_tray/tests/test_printing_tray.py
@@ -5,7 +5,7 @@
 from openerp.tests.common import TransactionCase
 
 
-model = 'openerp.addons.base_report_to_printer.printing'
+model = 'openerp.addons.base_report_to_printer.models.printing_server'
 
 
 class TestPrintingTray(TransactionCase):
@@ -13,8 +13,10 @@ class TestPrintingTray(TransactionCase):
     def setUp(self):
         super(TestPrintingTray, self).setUp()
         self.Model = self.env['printing.tray']
+        self.server = self.env['printing.server'].create({})
         self.printer = self.env['printing.printer'].create({
             'name': 'Printer',
+            'server_id': self.server.id,
             'system_name': 'Sys Name',
             'default': True,
             'status': 'unknown',
