+++ printer_tray/views/printing_printer.xml
